Made it use Dynamic Programming
Test with multiple money Systems to see if it works, but I think it does.
